refactor: create constants folder
